separate computer dependencies from Makefile unified version/spin locking safer thread stopping indication thread properties multi-producer/consumer circular buffering

+++ circ.h
@@ -1,9 +1,10 @@
-/* lock-less, single producer/consumer circular buffer */
+/* thread-safe circular buffer */
 
 #ifndef CIRC_H
 #define CIRC_H
 
 #include "status.h"
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -12,10 +13,10 @@ extern "C" {
 struct circ_t;
 typedef struct circ_t* circ_handle;
 
-status circ_create(circ_handle* pcirc, unsigned capacity);
+status circ_create(circ_handle* pcirc, size_t capacity);
 void circ_destroy(circ_handle* pcirc);
 
-unsigned circ_get_count(circ_handle circ);
+size_t circ_get_count(circ_handle circ);
 
 status circ_insert(circ_handle circ, void* val);
 status circ_remove(circ_handle circ, void** pval);
